refactor: enforce dupl threshold and extract shared provider shim helpers
