+++ .github/workflows/test.yml
@@ -130,6 +130,32 @@ jobs:
           only-new-issues: true
           args: --config=${{ github.workspace }}/.golangci.yml
 
+  provider-shim-checks:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
+        with:
+          fetch-depth: 0
+
+      - name: Setup Go
+        uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7.0.0
+        with:
+          go-version-file: providers/pkg/shim/go.mod
+          cache-dependency-path: providers/pkg/shim/go.sum
+
+      - name: Lint changed provider shim code
+        uses: golangci/golangci-lint-action@4afd733a84b1f43292c63897423277bb7f4313a9 # v8.0.0
+        with:
+          version: v2.7.2
+          working-directory: providers/pkg/shim
+          only-new-issues: true
+          args: --config=${{ github.workspace }}/.golangci.yml
+
+      - name: Test provider shim
+        run: go test -C providers/pkg/shim ./...
+
   actionlint:
     runs-on: ubuntu-latest
