+++ controller/internal/controller/modeldeployment_controller.go
@@ -248,10 +248,7 @@ func (r *ModelDeploymentReconciler) Reconcile(ctx context.Context, req ctrl.Requ
 	if r.EnableProviderSelector {
 		if err := r.selectEngine(ctx, &md, providerConfigs, resolvedServingMode); err != nil {
 			logger.Error(err, "Engine selection failed", "name", md.Name)
-			r.setCondition(&md, airunwayv1alpha1.ConditionTypeEngineSelected, metav1.ConditionFalse, "SelectionFailed", err.Error())
-			md.Status.Message = fmt.Sprintf("Engine selection failed: %s", err.Error())
-			r.recordReconcileError(&md, "engine_selection")
-			return ctrl.Result{}, r.Status().Patch(ctx, &md, client.MergeFrom(base))
+			return r.patchSelectionFailure(ctx, &md, base, airunwayv1alpha1.ConditionTypeEngineSelected, "Engine selection", "engine_selection", err)
 		}
 	}
 
@@ -308,10 +305,7 @@ func (r *ModelDeploymentReconciler) Reconcile(ctx context.Context, req ctrl.Requ
 	if r.EnableProviderSelector {
 		if err := r.selectProvider(ctx, &md, providerConfigs, resolvedEngineType, resolvedServingMode); err != nil {
 			logger.Error(err, "Provider selection failed", "name", md.Name)
-			r.setCondition(&md, airunwayv1alpha1.ConditionTypeProviderSelected, metav1.ConditionFalse, "SelectionFailed", err.Error())
-			md.Status.Message = fmt.Sprintf("Provider selection failed: %s", err.Error())
-			r.recordReconcileError(&md, "provider_selection")
-			return ctrl.Result{}, r.Status().Patch(ctx, &md, client.MergeFrom(base))
+			return r.patchSelectionFailure(ctx, &md, base, airunwayv1alpha1.ConditionTypeProviderSelected, "Provider selection", "provider_selection", err)
 		}
 	}
 
@@ -743,6 +737,13 @@ func (r *ModelDeploymentReconciler) setImageFieldConflictStatus(md *airunwayv1al
 	}
 }
 
+func (r *ModelDeploymentReconciler) patchSelectionFailure(ctx context.Context, md, base *airunwayv1alpha1.ModelDeployment, conditionType, messagePrefix, metricType string, err error) (ctrl.Result, error) {
+	r.setCondition(md, conditionType, metav1.ConditionFalse, "SelectionFailed", err.Error())
+	md.Status.Message = fmt.Sprintf("%s failed: %s", messagePrefix, err.Error())
+	r.recordReconcileError(md, metricType)
+	return ctrl.Result{}, r.Status().Patch(ctx, md, client.MergeFrom(base))
+}
+
 // setCondition updates a condition on the ModelDeployment.
 //
 // LastTransitionTime is passed as metav1.Now() here, but
@@ -881,31 +882,20 @@ func (r *ModelDeploymentReconciler) recordMetrics(md *airunwayv1alpha1.ModelDepl
 	// Update the phase cache and apply gauge deltas (decrement old, increment new).
 	entry.Phase = currentPhase
 	r.phaseCacheMu.Lock()
-	decrementPhaseEntryGauges(previous)
-	incrementPhaseEntryGauges(entry)
+	applyPhaseEntryGaugeDelta(previous, -1)
+	applyPhaseEntryGaugeDelta(entry, 1)
 	r.phaseCache[key] = entry
 	r.phaseCacheMu.Unlock()
 }
 
-// decrementPhaseEntryGauges subtracts a phaseEntry's contributions from the aggregate gauges.
-func decrementPhaseEntryGauges(e phaseEntry) {
+// applyPhaseEntryGaugeDelta adds a phaseEntry's signed contribution to aggregate gauges.
+func applyPhaseEntryGaugeDelta(e phaseEntry, delta float64) {
 	replicaStates := []string{"desired", "ready", "available"}
 	if e.Phase != "" {
-		airmetrics.DeploymentStatus.WithLabelValues(e.Provider, string(e.Phase)).Dec()
+		airmetrics.DeploymentStatus.WithLabelValues(e.Provider, string(e.Phase)).Add(delta)
 	}
 	for i, s := range replicaStates {
-		airmetrics.DeploymentReplicas.WithLabelValues(e.Provider, s).Sub(float64(e.Replicas[i]))
-	}
-}
-
-// incrementPhaseEntryGauges adds a phaseEntry's contributions to the aggregate gauges.
-func incrementPhaseEntryGauges(e phaseEntry) {
-	replicaStates := []string{"desired", "ready", "available"}
-	if e.Phase != "" {
-		airmetrics.DeploymentStatus.WithLabelValues(e.Provider, string(e.Phase)).Inc()
-	}
-	for i, s := range replicaStates {
-		airmetrics.DeploymentReplicas.WithLabelValues(e.Provider, s).Add(float64(e.Replicas[i]))
+		airmetrics.DeploymentReplicas.WithLabelValues(e.Provider, s).Add(float64(e.Replicas[i]) * delta)
 	}
 }
 
@@ -913,7 +903,7 @@ func incrementPhaseEntryGauges(e phaseEntry) {
 func (r *ModelDeploymentReconciler) cleanupMetrics(key k8stypes.NamespacedName) {
 	r.phaseCacheMu.Lock()
 	if old, ok := r.phaseCache[key]; ok {
-		decrementPhaseEntryGauges(old)
+		applyPhaseEntryGaugeDelta(old, -1)
 		delete(r.phaseCache, key)
 	}
 	r.phaseCacheMu.Unlock()
