+++ providers/dynamo/config.go
@@ -22,7 +22,6 @@ import (
 	"fmt"
 	"time"
 
-	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -32,6 +31,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	airunwayv1alpha1 "github.com/ai-runway/airunway/controller/api/v1alpha1"
+	"github.com/ai-runway/airunway/providers/pkg/shim"
 )
 
 const (
@@ -215,36 +215,8 @@ func (m *ProviderConfigManager) Register(ctx context.Context) error {
 		return fmt.Errorf("failed to build annotations: %w", err)
 	}
 
-	config := &airunwayv1alpha1.InferenceProviderConfig{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        ProviderConfigName,
-			Annotations: annotations,
-		},
-		Spec: GetProviderConfigSpec(),
-	}
-
-	existing := &airunwayv1alpha1.InferenceProviderConfig{}
-	err = m.client.Get(ctx, types.NamespacedName{Name: ProviderConfigName}, existing)
-
-	if errors.IsNotFound(err) {
-		logger.Info("Creating InferenceProviderConfig", "name", ProviderConfigName)
-		if err := m.client.Create(ctx, config); err != nil {
-			return fmt.Errorf("failed to create InferenceProviderConfig: %w", err)
-		}
-	} else if err != nil {
-		return fmt.Errorf("failed to get InferenceProviderConfig: %w", err)
-	} else {
-		existing.Spec = config.Spec
-		if existing.Annotations == nil {
-			existing.Annotations = make(map[string]string)
-		}
-		for k, v := range annotations {
-			existing.Annotations[k] = v
-		}
-		logger.Info("Updating InferenceProviderConfig", "name", ProviderConfigName)
-		if err := m.client.Update(ctx, existing); err != nil {
-			return fmt.Errorf("failed to update InferenceProviderConfig: %w", err)
-		}
+	if err := shim.RegisterProviderConfig(ctx, m.client, ProviderConfigName, annotations, GetProviderConfigSpec()); err != nil {
+		return err
 	}
 
 	// Update status — check if backend CRD is installed
