+++ providers/vllm/go.mod
@@ -4,6 +4,7 @@ go 1.25.3
 
 require (
 	github.com/ai-runway/airunway/controller v0.0.0
+	github.com/ai-runway/airunway/providers/pkg/shim v0.0.0
 	github.com/google/go-containerregistry v0.21.4
 	k8s.io/api v0.35.3
 	k8s.io/apimachinery v0.35.3
@@ -112,3 +113,5 @@ require (
 )
 
 replace github.com/ai-runway/airunway/controller => ../../controller
+
+replace github.com/ai-runway/airunway/providers/pkg/shim => ../pkg/shim
